Moves CScript to scripts/script.h.  Fixes more compile issues.
