+++ src/makefile.bsd
@@ -124,6 +124,7 @@ OBJS= \
     obj/timedata.o \
     obj/sync.o \
     obj/script/compressor.o \
+    obj/script/script.o \
     obj/script/scriptutils.o \
     obj/script/sign.o \
     obj/txmempool.o \
