+++ src/makefile.osx
@@ -94,6 +94,7 @@ OBJS= \
     obj/rpcsmessage.o \
     obj/timedata.o \
     obj/script/compressor.o \
+    obj/script/script.o \
     obj/script/scriptutils.o \
     obj/script/sign.o \
     obj/sync.o \
