+++ src/primitives/block.cpp
@@ -7,6 +7,7 @@
 #include "protocol.h"
 #include "primitives/block.h"
 #include "chainparams.h"
+#include "script/sign.h"
 #include <boost/foreach.hpp>
 
 bool CBlock::CheckBlockSignature() const
